requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-multipart>=0.0.6
